refactor: move src/composables tests into test/src/composables
Move all 20 composable test files from src/composables/ to test/src/composables/
and update import paths accordingly.
