+++ test/src/composables/useAppConfig.spec.ts
