+++ test/src/composables/useChatLauncher.spec.ts
