+++ test/src/composables/useDirConfig.spec.ts
