+++ test/src/composables/useFaviconState.spec.ts
