+++ test/src/composables/useNewTerminal.spec.ts
