+++ test/src/composables/useTerminalConnections.spec.ts
