+++ test/src/composables/useWikiBrowse.spec.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeAll, beforeEach } from "vitest";
 import { createApp, defineComponent } from "vue";
 import { flushPromises } from "@vue/test-utils";
-import { router } from "../router";
+import { router } from "../../src/router";
 import { useWikiBrowse, wikiGotoIndex, wikiGotoPage, wikiGotoGraph, wikiGotoLint, wikiRouteSlug, wikiClose } from "./useWikiBrowse";
 
 // Install the singleton router into a throwaway app so currentRoute tracks pushes.
